Resolve out of memory problem when predicting the data.
predict.sparsegl function would create large dense matrix via the cbind2(1, newx) %*% nbeta operation, which can lead to memory issues for large inputs.

One way to mitigate memory issues is to compute the matrix-vector product in chunks, rather than all at once. The idea is to break the matrix and/or the vector into smaller chunks, compute the product for each chunk, and then combine the results. This approach requires less memory than computing the product for the whole matrix and vector at once, though it might be somewhat slower due to the overhead of managing the chunks.